tests: Allow running slow tests via env var
The slow tests can be run via "pytest --slow" when running the tests
from the passari root directory, but if running the tests from upper
level (e.g. from passari-compose project which has passari as a
submodule) the "--test" option flag is not registered and cannot be
used.

Allow running the slow tests even in that case with a command line like:

    RUN_SLOW_TESTS=1 pytest

I.e. by setting environment variable RUN_SLOW_TESTS to 1.