WIP: Add DEBUG prints to MuseumPlus PUT request
This request used to fail when the field name was not set correctly, but
even after setting the correct(?) field name it still doesn't seem to
save the data to MuseumPlus.  The response to the PUT request was HTTP
code 204 (with no content) so it should have succeeded, but still the
field (`ObjPASLog01Clb`) seems to be empty on MuseumPlus.