Migrate to dpres-siptools-ng
Create the SIPs using dpres-siptools-ng instead of the command-line
toolset of dpres-siptools. The stucture map and technical metadata
of files in the SIP are generated automatically by dpres-siptools-ng.

There are a couple of notable changes:
- Update SIP creationg is temporarily not possible.
- The folder structure of the compressed SIP changes so that
the mets.xml and signature.sig remain in the root but the source
files are in a sub-directory. This is how dpres-siptools-ng
generates the SIP.
- Passari had some custom PreservationError detectors designed
around the import-object command line script provided by dpres-siptools.
These are no longer compatible. The problematic image files that
raised such errors are still rejected when creating the SIP but
the error messages are different.

Refs MPZ-5